test(bookmark): spike3 passes simple args passing...
Green with a TODO: test. Next, test dashdash args. I made the
test data and helper functions already in /examples/simple2/tests/
testUtils.ts and testData.ts. See the cmd spike 3 tests for examples

tests: green
bookmark: /experiments/simple2/spikes/cmd_parser_3.test.ts
